+++ src/main/java/cech12/unlitcampfire/mixinaccess/ICampfireBlockEntityMixin.java
@@ -1,0 +1,13 @@
+package cech12.unlitcampfire.mixinaccess;
+
+public interface ICampfireBlockEntityMixin {
+
+    boolean isSoulCampfire();
+
+    int getLitTime();
+
+    boolean addLitTime(int litTimeToAdd);
+
+    boolean removeLitTime(int litTimeToRemove);
+
+}